fix for size

+++ Human/src/Rendering/OPshader.cpp
@@ -6,9 +6,9 @@
 #include "./Core/include/OPlog.h"
 
 //-----------------------------------------------------------------------------
-// ______                _   _                 
-//|  ____|              | | (_)                
-//| |__ _   _ _ __   ___| |_ _  ___  _ __  ___ 
+// ______                _   _
+//|  ____|              | | (_)
+//| |__ _   _ _ __   ___| |_ _  ___  _ __  ___
 //|  __| | | | '_ \ / __| __| |/ _ \| '_ \/ __|
 //| |  | |_| | | | | (__| |_| | (_) | | | \__ \
 //|_|   \__,_|_| |_|\___|\__|_|\___/|_| |_|___/
@@ -25,7 +25,7 @@ OPint OPshaderLoadVertex(OPstream* source, OPshader** shader){
 		OPchar* src = (OPchar*)source->Data;
 
 		GLint lengths[1] = {
-			source->Size
+			(GLint)source->Size
 		};
 		glShaderSource(vertex, 1, (const OPchar**)&src, lengths);
 		OPglError("GLShader::Error 2");
@@ -72,7 +72,7 @@ OPint OPshaderLoadFragment(OPstream* source, OPshader** shader){
 	if (frag){
 		OPchar* src = (OPchar*)source->Data;
 		GLint lengths[1] = {
-			source->Size
+			(GLint)source->Size
 		};
 		glShaderSource(frag, 1, (const OPchar**)&src, lengths);
 		glCompileShader(frag);
